# artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyFortress",
    "version": "0.1.0"
  },
  "paths": {
    "/home-assistant/probe": {
      "post": {
        "summary": "Probe Home Merlin entities",
        "operationId": "home_assistant_probe_home_assistant_probe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HomeMerlinProbeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeMerlinProbeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/home-assistant/service": {
      "post": {
        "summary": "Invoke Home Merlin service",
        "operationId": "home_assistant_service_home_assistant_service_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HomeMerlinServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeMerlinServiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/home-assistant/state": {
      "post": {
        "summary": "Set Home Merlin entity state",
        "operationId": "home_assistant_state_home_assistant_state_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HomeMerlinStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeMerlinStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/frigate/probe": {
      "post": {
        "summary": "Probe Frigate version",
        "operationId": "frigate_probe_frigate_probe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FrigateProbeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrigateProbeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/frigate/cameras": {
      "get": {
        "summary": "List Frigate cameras from config",
        "operationId": "frigate_cameras_frigate_cameras_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrigateCamerasResponse"
                }
              }
            }
          }
        }
      }
    },
    "/frigate/events": {
      "get": {
        "summary": "List Frigate events",
        "operationId": "frigate_events_frigate_events_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrigateEventsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/frigate/events/stream": {
      "get": {
        "summary": "Stream Frigate events (Server-Sent Events)",
        "operationId": "frigate_events_stream_frigate_events_stream_get",
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        }
      }
    },
    "/snapshot": {
      "post": {
        "summary": "Aggregate snapshot across Home Merlin and Frigate",
        "operationId": "snapshot_snapshot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/system/info": {
      "get": {
        "summary": "System Info",
        "operationId": "system_info_system_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response System Info System Info Get"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/readiness": {
      "get": {
        "summary": "Readiness",
        "operationId": "readiness_readiness_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Readiness Readiness Get"
                }
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Metrics Endpoint",
        "operationId": "metrics_endpoint_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Metrics Endpoint Metrics Get"
                }
              }
            }
          }
        }
      }
    },
    "/metrics/prom": {
      "get": {
        "summary": "Metrics Prometheus",
        "operationId": "metrics_prometheus_metrics_prom_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Metrics Prometheus Metrics Prom Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EntityReading": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "state": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "entity_id"
        ],
        "title": "EntityReading"
      },
      "FrigateCamerasResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "cameras": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cameras"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "FrigateCamerasResponse"
      },
      "FrigateEventsResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Events"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "stream_supported": {
            "type": "boolean",
            "title": "Stream Supported",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "FrigateEventsResponse"
      },
      "FrigateProbeRequest": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "Frigate base URL"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "API key"
          },
          "cameras": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cameras"
          }
        },
        "type": "object",
        "title": "FrigateProbeRequest"
      },
      "FrigateProbeResponse": {
        "properties": {
          "healthy": {
            "type": "boolean",
            "title": "Healthy"
          },
          "version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FrigateVersion"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cameras": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cameras"
          }
        },
        "type": "object",
        "required": [
          "healthy"
        ],
        "title": "FrigateProbeResponse",
        "description": "Response for Frigate probe."
      },
      "FrigateSnapshot": {
        "properties": {
          "healthy": {
            "type": "boolean",
            "title": "Healthy"
          },
          "version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FrigateVersion"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cameras": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cameras"
          }
        },
        "type": "object",
        "required": [
          "healthy"
        ],
        "title": "FrigateSnapshot"
      },
      "FrigateVersion": {
        "properties": {
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra"
          }
        },
        "type": "object",
        "title": "FrigateVersion"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HomeMerlinProbeRequest": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "Home Merlin base URL"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token",
            "description": "Long-lived token"
          },
          "entities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entities"
          },
          "verify_ssl": {
            "type": "boolean",
            "title": "Verify Ssl",
            "default": true
          }
        },
        "type": "object",
        "title": "HomeMerlinProbeRequest"
      },
      "HomeMerlinProbeResponse": {
        "properties": {
          "healthy": {
            "type": "boolean",
            "title": "Healthy"
          },
          "readings": {
            "additionalProperties": {
              "$ref": "#/components/schemas/EntityReading"
            },
            "type": "object",
            "title": "Readings"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "healthy"
        ],
        "title": "HomeMerlinProbeResponse",
        "description": "Response for Home Merlin probe."
      },
      "HomeMerlinServiceRequest": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "domain",
          "service"
        ],
        "title": "HomeMerlinServiceRequest",
        "description": "Invoke a Home Merlin service (domain/service)."
      },
      "HomeMerlinServiceResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "response": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "HomeMerlinServiceResponse"
      },
      "HomeMerlinSnapshot": {
        "properties": {
          "healthy": {
            "type": "boolean",
            "title": "Healthy"
          },
          "readings": {
            "additionalProperties": {
              "$ref": "#/components/schemas/EntityReading"
            },
            "type": "object",
            "title": "Readings"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "healthy"
        ],
        "title": "HomeMerlinSnapshot"
      },
      "HomeMerlinStateRequest": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "state": {
            "title": "State"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          }
        },
        "type": "object",
        "required": [
          "entity_id",
          "state"
        ],
        "title": "HomeMerlinStateRequest"
      },
      "HomeMerlinStateResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "state": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "HomeMerlinStateResponse"
      },
      "SnapshotRequest": {
        "properties": {
          "include_home_assistant": {
            "type": "boolean",
            "title": "Include Home Merlin",
            "default": true
          },
          "include_frigate": {
            "type": "boolean",
            "title": "Include Frigate",
            "default": true
          },
          "home_assistant_entities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Home Merlin Entities"
          },
          "include_frigate_cameras": {
            "type": "boolean",
            "title": "Include Frigate Cameras",
            "description": "Include camera list",
            "default": true
          }
        },
        "type": "object",
        "title": "SnapshotRequest"
      },
      "SnapshotResponse": {
        "properties": {
          "home_assistant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HomeMerlinSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "frigate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FrigateSnapshot"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "SnapshotResponse",
        "description": "Aggregated snapshot response."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}